Change name `singularity` to `apptainer`
