+++ repo2apptainer/__init__.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 # flake8: noqa
-""" Top-level module for repo2singularity. """
+""" Top-level module for repo2apptainer. """
 from ._version import version as __version__
-from .app import Repo2Singularity
+from .app import Repo2Apptainer
 from .config import config
