+++ repo2apptainer/cli.py
@@ -1,0 +1,43 @@
+import typer
+
+from ._version import version as __version__
+from .app import Repo2Apptainer
+from .console import console
+
+app = typer.Typer(help='Convert a repository to a Apptainer (formerly Singularity) image')
+
+
+def version_callback(value: bool):
+    if value:
+        typer.echo(f'Repo2Apptainer Version: {__version__}')
+        raise typer.Exit()
+
+
+@app.command()
+def build(repo: str, ref: str = 'main', force: bool = False):
+    """Build an Apptainer/Singularity image from a repository"""
+    console.print('Building an Apptainer/Singularity image from a repository')
+    r2s = Repo2Apptainer(repo=repo, ref=ref, force=force)
+    console.print(f'Repository: {r2s.r2d.repo}')
+    console.print(f'Reference: {r2s.r2d.ref}')
+    console.print(f'Output image: {r2s.r2d.output_image_spec}')
+    console.print(f'Cache directory: {r2s.cache_dir}')
+    r2s.build_docker()
+    r2s.build_sif()
+    console.print(f'SIF image: {r2s.sif_image}')
+
+
+@app.command()
+def run():
+    """Run an Apptainer/Singularity image"""
+    console.print('Running an Apptainer/Singularity image')
+
+
+@app.command()
+def version():
+    """Print the version"""
+    console.print(f'Repo2Apptainer version: {__version__}')
+
+
+def main():
+    typer.run(app())
